requirements.txt
google-cloud-storage
google-cloud-translate
python-docx
